Rating: Rename RatingBase.tsx to Rating.base.tsx to match convention (#3921)
* Rename RatingBase.tsx to Rating.base.tsx to match convention

* Add change file
