+++ common/changes/office-ui-fabric-react/miwhea-rename-rating-base_2018-02-08-16-38.json
@@ -1,0 +1,11 @@
+{
+  "changes": [
+    {
+      "packageName": "office-ui-fabric-react",
+      "comment": "Rating: Renamed RatingBase.tsx to Rating.base.tsx to match convention",
+      "type": "patch"
+    }
+  ],
+  "packageName": "office-ui-fabric-react",
+  "email": "mike@mikewheaton.ca"
+}
