+++ packages/office-ui-fabric-react/src/components/Rating/Rating.base.tsx
